refactor(ui): split tour overlay component into modular files
BREAKING CHANGE: tour_overlay.dart has been removed and replaced with
multiple dedicated files under tour_overlay/ directory. Update imports
from 'package:.../tour_overlay.dart' to
'package:.../tour_overlay/tour_overlay.dart'
