+++ src/lib/presentation/ui/features/calendar/pages/today_page.dart
@@ -33,7 +33,7 @@ import 'package:whph/presentation/ui/shared/services/abstraction/i_confetti_anim
 import 'package:whph/presentation/ui/shared/services/abstraction/i_translation_service.dart';
 import 'package:whph/presentation/ui/shared/services/abstraction/i_theme_service.dart';
 import 'package:acore/utils/responsive_dialog_helper.dart';
-import 'package:whph/presentation/ui/shared/components/tour_overlay.dart';
+import 'package:whph/presentation/ui/shared/components/tour_overlay/tour_overlay.dart';
 import 'package:whph/presentation/ui/shared/services/tour_navigation_service.dart';
 import 'package:whph/presentation/ui/features/calendar/constants/calendar_translation_keys.dart';
 import 'package:whph/presentation/ui/shared/components/section_header.dart';
